src/requirements.txt
beautifulsoup4==4.8.1
channels==3.0.4
channels-redis==3.3.1
Django==4.0.2
django-cleanup==6.0.0
django-debug-toolbar==3.2.4
django-extensions==3.1.5
django-otp==1.1.3
django-qr-code==1.1.0
django-redis==5.2.0
fastcore==1.3.27
pdfkit==1.0.0
Pillow==7.1.2
python-dotenv==0.14.0
python-social-auth==0.3.6
qrcode==6.1
redis==4.1.2
requests==2.27.1
rich==11.2.0
social-auth-app-django==5.0.0
social-auth-core==4.2.0
tweepy==4.5.0
gunicorn==20.1.0